updated main

+++ src/main.cpp
@@ -32,12 +32,12 @@ int main() {
     time_t t = time(0);
     struct tm *now = localtime(&t);
     
-    double year = now->tm_year + 1900;
-    double month = now->tm_mon + 1;
-    double day = now->tm_mday;
+    int year = now->tm_year + 1900;
+    int month = now->tm_mon + 1;
+    int day = now->tm_mday;
 
-    double daysInMonth = getDaysInMonth(month, year);
-    double remainingDays = daysInMonth - day;
+    int daysInMonth = getDaysInMonth(month, year);
+    int remainingDays = daysInMonth - day;
 
     int choice;
     clearScreen();
@@ -86,7 +86,7 @@ int main() {
                 std::cout << "\nTotal: " << manager.getSum() << "\n" << std::endl;
                 double diff = goal - manager.getSum();
                 if (diff > 0){
-                    double remaining = diff / remainingDays;
+                    int remaining = diff / remainingDays;
                     std::cout << "\nRemaining in budget:" << diff << "\n" << std::endl;
                     std::cout << "You can spend " << remaining << " per day for the remainder of the month\n" << std::endl;
                 }
